Improved cleanliness of publicly shwon URLs for each view when the name is "index" -- in which case this will always be displayed as the empth string; also supplied a VersionLogView.action_url() method that returns an empty string (otherwise forms submitted with post will get a "@@index" appended to their URL).
Changed form action default to the more correct empty string (instead of "." - note: this is anyway overwritten by the form's action_url() or the request's URL).

Un-hardwired the form submit method (that was always "post") to the (so far dummy) form.action_method() utility -- for the cases when only information is retrieved (i.e. no domain objects are modified) this should be "get".