Schema changes needed for implementing Issue 612 and Issue 662: - effects the 6 *_changes tables: agenda_item_changes, attached_file_changes, bill_changes, motion_changes, question_changes, tabled_document_changes - each of these 6 tables has been modified by changing the single "date" column to the 2 columns "date_audit" and "date_active", both of type "timestamp without time zone" and "not null" (note the origininal audit "date" column was previously nullable).
Existing databases needed to be updated.

The date_audit column is exclusively managed by the system.

The date_active column is user-modifiable, defaulting to the same value as date_audit, 
and is the *effective* date i.e. the date to be used for all intents and purposes other than for data auditing.

The UI has been updated as follows:
- all workflow views: the "date" column nows shows "date_active";
- all versions views: the "modified" column now shows "date_audit";
- all audit-log views: the "date" column now shows "date_active"; a newly added view column "audit date" renders the value of "date_audit";
- all timeline tabs: the "date" column now shows "date_active";

Note currently newly added date_active columns are not as yet exposed for editing anywhere in the UI.