Parliamentary Items Review:
SCHEMA CHANGE: deleted column "notice_date" on the db table "motions" (that was never being set anywhere in the code) and replaced it with the property "notice_date" on domain.Motion that will extract this date dynamically from the workflow audit log of the motion, as the date that it was "scheduled". 

Adjusted MotionDescriptor field for this to not be editable, and to be displayed in (view, listing) allowing localization.