[naming, type_info] Adjusting the UI RSSValues adapter to retrieve the workflow via domain_mdoel directly, not via obsoleted get_workflow and incorrect naming convention; fixes resported probelm with download of RSS for TabledDocuments.
Adding a bool custom to TypeInfo, to keep track of whether the type was loaded from custom conf or not (so system type). Using it to Unhardwire the list of user defined types in the ui.forms.viewlets.MemberItemsViewlet; replacing (with capi.get_type_info) other usage of get_workflow in same module.