[descriptor] Switching following descriptors to use new field declaration format:     TitleTypeDescriptor     MemberTitleDescriptor     PoliticalGroupDescriptor     OfficeDescriptor     MinistryDescriptor     GovernmentDescriptor     AttachmentDescriptor
Adding support for render_types "no_input" and "file". 

Changing name of vocabulary class TitleTypes to GroupTitleTypesFactory.

registering vocabularies: 
    group_sub_role_factory as "group_sub_role"
    office_role_factory as "office_role"
    GroupTitleTypesFactory() as "group_title_types"

Changing the AttachmentDescriptor fields "name" and "mimetype" to be displayed in also in "view" and "listing" (!+ latter does not seem to work anyway).

Correcting base GroupDescriptor field "end_date" to be not required.