[member addresses cleanout] Cleaning out viewlet, add menu item, etc after removal of member.addresses container property in r10480; corrects the observed problem to "add a user as mp".
Adding "titles" container property to political_group_member, that was lost in the refactor of these in r10480; corrects observed problem in "adding a title to a political group member".

Some refinement of revisit comments on the files/events container properties on Version classes.

Tweaking whitespace/indentation production of default forms/ui.xml.