[types, descriptor] Adding support for a @descriptor attribute on declaration of custom types, allowing direct re-use of a descriptor.
Adding a descriptor_key attr to TypeInfo, that remembers the types @descriptor value, then to be evaluated either as the file name for a descriptor declaration or as the key for a previously loaded descriptor. Dropping @name attr on declaration of custom (one-per-file) descriptors; always the name of the file.

Extending pattern of using custom archetypes (introduced in r10931) for all "equivalent" doc types in Assembly and in Senate; re-using doc descriptor (and a placeholder doc workflow) for each of the "intermediate" types 9the custom archetypes).

Adding type_key as an attr on TypeInfo; reworking type registry internal data structure, to now simply be a list of TypeInfo instances, instead of a list of (key, instance) as previous, now the key bieng iyself an attr on the instance.

Couple minor corrections/tweaks in capi.