[code cleaning] Replacing the few odd log.warning() calls with log.warn() to be consistent across the code.
Dropping the use of the % string interpolation operator, replacing with comma-sparating the paramaters -- this has performance and possibly correctness implications [usage of % operator will cause all params to be executed, while letting the logging machinery do the necessary interpolation below (via usage of comma to separate params) these params need not be executed if logging level does not include the level we are logging at.

Correcting typo in logging message (from r11525) for "Doc.oncreate" that should be "Doc.on_create" (as per the method name).

Dropping the call to dict.keys() in inclusion check (emailnotifications) -- this is not necessary to check whether a dict has a certain entry, and again, has performance implications.